fixing analysis and documentations adding new tests

+++ analyses/conversion_cost.sql
@@ -1,6 +1,6 @@
 select 
         channel, 
-        sum(pa.spend)/nullif(sum(pa.total_conversions),0)::float  as cost_conversion
+        sum(pa.spend)/nullif(sum(pa.total_conversions),0)  as cost_conversion
     from {{ ref('paid_ads') }} as pa 
     
 group by channel
