+++ analyses/cost_engage.sql
@@ -1,6 +1,6 @@
 select 
         channel, 
-        sum(pa.spend)/nullif(sum(pa.engagements),0)::float  as cost_engage
+        sum(pa.spend)/nullif(sum(pa.engagements),0)  as cost_engage
     from {{ ref('paid_ads') }} as pa 
 group by channel
